Initial Commit
homescreen.dart: serves as the main screen. Composed of Map widget, AppBar(with search) widget, filter button, and temporary bottom menu.

bottom_sheet.dart: serves as the temporary bottom sheet widget that will list all events, implementation of the sheet likely will remain similar to the current, but usage of it will differ

event_card.dart: a card widget customized to hold events as they should be displayed in the bottom_sheet screen. This is entirely based off of the Adobe XD draft

event_object.dart: a class with the goal of entirely containing all of the information necessary to list, describe, and filter the event. Ideally, the database structure mirrors this, or this mirrors the database structure

event_popup.dart: a widget that pops up when you select an event from the BottomSheetWidget

filtersheet.dart: A right-sided drawer that slides in when the action button on the appbar is tapped. Contains all filters that apply to the event pool

main.dart: establishes a base MaterialApp (to avoid styling issues... the homescreen makes a new one too) and immediately launches into the HomeScreen.

TODO:
1. Add login screen
2. Finish event_popup.dart
3. Implement event_card into bottom_sheet
4. Link filtersheet with the events listed
5. Create a large-scope collection of all of the events (maybe as a stream so we can listen for new events)

Other items:
1. Mock data (ideally on Firebase)
  a. If we can start to link Firebase to the app, we can see how they will interact (and improve our schema)
2. Profile page/ algorithms for preferences
3. Write test backend functions for managing each user